+++ CHANGES.md
@@ -7,6 +7,8 @@
   variable.
 - [FIXED] Fix `get_security`/`set_security` asymmetry.
 - [FIXED] Support piping of request payload.
+- [BREAKING CHANGE] Rename client plugin configuration flags:
+  `retryAttempts` -> `maxAttempt` & `retryTimeout` -> `retryInitialDelayMsecs`
 - [IMPROVED] Updated documentation by replacing deprecated Cloudant links with
   the latest bluemix.net links.
 - [REMOVED] Remove previously deprecated method `set_permissions`.